#trend chart

+++ OEE-API/ViewModels/ChartTrendViewModel.cs
@@ -5,6 +5,6 @@ namespace OEE_API.ViewModels
     public class ChartTrendViewModel
     {
         public string name { get; set; }
-        public List<int> data { get; set; }
+        public List<decimal> data { get; set; }
     }
 }
